+++ src/nips/mod.rs
@@ -5,4 +5,5 @@ pub mod nip16;
 pub mod nip2;
 pub mod nip25;
 pub mod nip4;
+pub mod nip5;
 pub mod nip9;